[BAD CODE, NOT WORKING] First attempts to implement class constuctor in c++
This code is kinda bad because it does not work lol ahahahah. It is 1:30 am now and i do not want to go watch lectures all over again to remember something about c++ after my 3 days Rust adventure. So now ima gonna sleep and laugh at this code a couple of days later.